Refactor split module, fix BLE split
